+++ aiogram/types/chat_member_union.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
-from typing import TypeAlias
+from typing import Annotated, TypeAlias
+
+from pydantic import Field
 
 from .chat_member_administrator import ChatMemberAdministrator
 from .chat_member_banned import ChatMemberBanned
@@ -9,11 +11,12 @@ from .chat_member_member import ChatMemberMember
 from .chat_member_owner import ChatMemberOwner
 from .chat_member_restricted import ChatMemberRestricted
 
-ChatMemberUnion: TypeAlias = (
+ChatMemberUnion: TypeAlias = Annotated[
     ChatMemberOwner
     | ChatMemberAdministrator
     | ChatMemberMember
     | ChatMemberRestricted
     | ChatMemberLeft
-    | ChatMemberBanned
-)
+    | ChatMemberBanned,
+    Field(discriminator="status"),
+]